requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container run 'pip freeze > requirements.txt' in the terminal to overwrite this file
pywhatkit
pygame
Pandas
NumPy
matplotlib
opencv-python
pywhatkit
opencv-contrib-python
youtube-dl
PyAutoGUI
mysql-connector-python
imutils
pyyaml
img2pdf
psutil
openai
pyatv
flask-socketio
Flask
geopy
selenium
chromedriver_autoinstaller
Panda3D
mplfinance
keyboard
Pillow
